# docs/swagger.yaml
definitions:
  helpers.HttpDefaultResponse:
    properties:
      message:
        type: string
    type: object
...
  requests.UserLoginDto:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and retrieve a JWT for API calls
      operationId: Authentication
      parameters:
      - description: Credentials
        in: body
        name: '{object}'
        required: true
        schema:
          $ref: '#/definitions/requests.UserLoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.HttpDefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HttpDefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.HttpDefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HttpDefaultResponse'
      summary: Provides a JSON Web Token
...
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"